+++ lib/ServerlessProject.js
@@ -403,24 +403,6 @@ class ServerlessProject {
       })
   }
 
-  _filterSPaths( collection, options ) {
-    // TODO: remove this, caller should be filtering such stuff later
-    let paths = _.get( options, 'paths' );
-    if( paths && (paths.length > 0) ) {
-      collection = _.filter( collection, f =>
-        _.any( paths, p =>
-          p.indexOf( f.getSPath() ) !== -1;
-        )
-      )
-    }
-
-    if( _.get( options, 'returnPaths' ) == true ){
-      collection = _.map( collection, func => func.getSPath(); );
-    }
-
-    return collection;
-  }
-
   /**
    * Get Functions
    * - Returns an array of this project's function instances
@@ -430,7 +412,7 @@ class ServerlessProject {
    */
 
   getAllFunctions(options) {
-    return this._filterSPaths( _.flatten( _.map( _.values( this.components ), component =>
+    return SUtils.filterSPaths( _.flatten( _.map( _.values( this.components ), component =>
       component.getAllFunctions();
     )), options );
   }
@@ -449,7 +431,7 @@ class ServerlessProject {
    */
 
   getAllComponents(options){
-    return this._filterSPaths( this.components, options );
+    return SUtils.filterSPaths( this.components, options );
   }
 
   getComponent( path ){
@@ -466,7 +448,7 @@ class ServerlessProject {
    */
 
   getEvents(options) {
-    return this._filterSPaths( _.flatten( _.map( this.getAllFunctions(), f => f.getAllEvents() ) ), options );
+    return SUtils.filterSPaths( _.flatten( _.map( this.getAllFunctions(), f => f.getAllEvents() ) ), options );
   }
 
   getEvent( path ){
