+++ lib/actions/FunctionDeploy.js
@@ -204,7 +204,7 @@ module.exports = function(SPlugin, serverlessPath) {
 
       // If --all is selected, load all paths
       if (_this.evt.options.all) {
-        _this.evt.options.paths = _this.S.state.getFunctions({ returnPaths: true });
+        _this.evt.options.paths = _this.S.getProject().getAllFunctions({ returnPaths: true });
       }
 
       // Ensure tmp folder exists in _meta
@@ -269,7 +269,7 @@ module.exports = function(SPlugin, serverlessPath) {
 
           return BbPromise.try(function() {
 
-              func = _this.S.state.getFunctions({ paths: [path] })[0];
+              func = _this.S.getProject().getFunction( path );
               if (!func) throw new SError(`Function could not be found: ${path}`);
 
               let newEvt = {
