+++ lib/actions/PluginCreate.js
@@ -155,7 +155,7 @@ usage: serverless plugin create <plugin>`,
 
       // TODO: Remove in V1 because will result in breaking change
       // Add the newly create plugin to the plugins array of the projects s-project.json file
-      // TODO: this belongs to Project class
+      // TODO: this belongs to Project class XXX
       let sProjectJson = SUtils.readAndParseJsonSync(this.S.getProject().getFilePath('s-project.json'));
       sProjectJson.plugins.push(pluginName);
       fs.writeFileSync(this.S.getProject().getFilePath('s-project.json'), JSON.stringify(sProjectJson, null, 2));
