+++ lib/utils/index.js
@@ -16,6 +16,7 @@ let BbPromise   = require('bluebird'),
   SCli          = require('./cli'),
   fs            = require('fs'),
   mkdirpAsync   = require('mkdirp-then'),
+  _             = require('lodash'),
   shortid       = require('shortid');
 
 BbPromise.promisifyAll(fs);
@@ -557,6 +558,22 @@ exports.populate = function(meta, templates, data, stage, region) {
   return data;
 };
 
+exports.filterSPaths = function( collection, options ) {
+  let paths = _.get( options, 'paths' );
+  if( paths && (paths.length > 0) ) {
+    collection = _.filter( collection, f =>
+      _.any( paths, p =>
+        p.indexOf( f.getSPath() ) !== -1;
+      )
+    )
+  }
+
+  if( _.get( options, 'returnPaths' ) == true ){
+    collection = _.map( collection, func => func.getSPath(); );
+  }
+
+  return collection;
+};
 
 function pathToContext(path) {
   // Match files under lib, tests, or bin so we only report the
